refactor from updog to updog2
