+++ updog2/__init__.py
@@ -1,4 +1,4 @@
-version_info = (1,5)
+version_info = (1,0)
 version = '.'.join(str(c) for c in version_info)
 
 base_directory = ''
