+++ updog2/templates/home.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <!-- Title -->
-    <title>updog - {{ directory }}</title>
+    <title>Updog2 - {{ directory }}</title>
     <!-- Stylesheets -->
     <link rel="stylesheet" href="{{ url_for('static', filename='css/bootstrap.min.css') }}">
     <link rel="stylesheet" href="{{ url_for('static', filename='vendor/fontawesome-5.1/css/all.min.css') }}">
@@ -20,7 +20,7 @@
     <!-- Header -->
     <header id="header" class="d-flex align_item_center">
         <div onclick="document.location='/'" class="logo_p">
-            <img src="{{ url_for('static', filename='images/updog.png') }}" alt="updog">
+            <img src="{{ url_for('static', filename='images/updog2.png') }}" alt="updog2">
         </div>
 
         <div class="heading_title_p">
@@ -114,7 +114,7 @@
 
     <footer>
         <p>
-            updog v{{ version }}
+            updog2 v{{ version }}
         </p>
     </footer>
